+++ pages/portfolio/src/modals/transfer/TransferModalContent.tsx
@@ -3,7 +3,7 @@
 
 "use client";
 
-import { useEffect, useMemo } from "react";
+import { useEffect, useMemo, useState } from "react";
 import {
   ErrorMessage,
   IconContainer,
@@ -56,6 +56,7 @@ import { getTokenValidDestinations } from "../shared/getTokenValidDestinations";
 import { TransactionInspector } from "../shared/TransactionInspector";
 import { useTranslation } from "@evmosapps/i18n/client";
 import { ConnectToWalletWarning } from "../shared/ConnectToWalletWarning";
+import { normalizeToCosmos } from "helpers/src/crypto/addresses/normalize-to-cosmos";
 
 export const TransferModalContent = ({
   receiver,
@@ -112,7 +113,10 @@ export const TransferModalContent = ({
 
   const token = getTokenByRef(tokenRef);
   const senderChain = sender ? getChainByAddress(sender) : getChain("evmos");
-
+  const [receiverNetworkPrefix, setReceiverNetworkPrefix] = useState("evmos");
+  const receiverChain = receiver
+    ? getChainByAddress(receiver)
+    : getChain("evmos");
   const destinationNetworkOptions = useMemo(
     (): string[] => //
       getTokenValidDestinations(tokenAmount.ref, senderChain.prefix),
@@ -198,6 +202,14 @@ export const TransferModalContent = ({
     )
       return "TOPUP";
 
+    if (
+      token.sourcePrefix === "axelar" &&
+      token.handledByExternalUI !== null &&
+      networkPrefix === "evmos" &&
+      receiverNetworkPrefix === "evmos" &&
+      receiverChain.name === "Evmos"
+    )
+      return "TRANSFER";
     if (token.handledByExternalUI !== null) return "BRIDGE";
 
     return "TRANSFER";
@@ -209,8 +221,11 @@ export const TransferModalContent = ({
     validation.hasSufficientBalanceForFee,
     isPreparing,
     fee,
+    networkPrefix,
+    receiverNetworkPrefix,
+    token.sourcePrefix,
+    receiverChain.name,
   ]);
-
   useEffect(() => {
     if (!validation.hasSufficientBalanceForFee && !isPreparing)
       sendEvent(ERROR_IN_SEND, {
@@ -404,6 +419,7 @@ export const TransferModalContent = ({
               }
               networkOptions={destinationNetworkOptions}
               senderPrefix={senderChain.prefix}
+              setNetworkState={setReceiverNetworkPrefix}
             />
             {sender && receiver && amount !== 0n && (
               <div className="space-y-3 mt-8">
@@ -432,6 +448,13 @@ export const TransferModalContent = ({
                 {feeToken?.symbol}
               </ErrorMessage>
             )}
+
+            {receiver !== undefined &&
+              sender === normalizeToCosmos(receiver) && (
+                <ErrorMessage className="justify-center pl-0">
+                  {t("error.duplicatedAddress")}
+                </ErrorMessage>
+              )}
 
             {action === "TOPUP" && (
               <PrimaryButton